+++ Release_Notes.md
@@ -27,6 +27,8 @@
         Excluding:
                 - DataRow GetChoices(string account)
     - DAccounts
+    - DSecurityEntities
+        - Task DeleteRegistrationInformation(int securityEntitySeqId)
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be, causing a logout to occure